feat: connect worry timeline report API

+++ app/src/main/java/com/gominitta/android/data/remote/dto/AnxietyGapResponse.kt
@@ -5,9 +5,9 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class AnxietyGapResponse(
     val period: String,
-    val beforeScore: Double,
-    val afterScore: Double,
-    val gap: Double,
-    val sampleCount: Int,
+    val beforeScore: Long,
+    val afterScore: Long,
+    val gap: Long,
+    val sampleCount: Long,
     val feedback: String,
 )
